Rename abstract.core->abstract.loop
